1.4

+++ src/main/java/dev/lightdream/royalsecurity/managers/DiscordEventManager.java
@@ -1,7 +1,7 @@
 package dev.lightdream.royalsecurity.managers;
 
 import dev.lightdream.royalsecurity.Main;
-import dev.lightdream.royalsecurity.dto.User;
+import dev.lightdream.royalsecurity.database.User;
 import lombok.Getter;
 import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.events.interaction.ButtonClickEvent;
@@ -38,8 +38,22 @@ public class DiscordEventManager extends ListenerAdapter {
 
             user.setIP(ip);
         } else if (event.getComponentId().contains("deny_authentication")) {
+            String data = event.getComponentId().replace("deny_authentication_", "");
+            String ip = data.split(";")[0];
+            String name = data.split(";")[1];
+
+            event.getChannel().sendMessageEmbeds(plugin.jdaConfig.accessGranted.build().build()).queue();
+            event.getMessage().delete().queue();
+            User user = plugin.databaseManager.getUser(name);
+
+            if (user == null) {
+                event.getChannel().sendMessageEmbeds(Main.instance.jdaConfig.invalidUser.build().build()).queue();
+                return;
+            }
+
             event.getChannel().sendMessageEmbeds(plugin.jdaConfig.accessDenied.build().build()).queue();
             event.getMessage().delete().queue();
+
         }
     }
